backend/requirements.txt
fastapi
uvicorn
google-cloud-storage
slack_sdk
requests
pydantic-settings # Pydantic v2 needs this for BaseSettings
python-crontab # For cron string validation/parsing if needed
python-dotenv # For loading .env file